bulkdeliver(): Rework the SMTP error handling to fix a problem described here:
http://mail.python.org/pipermail/mailman-developers/2003-November/016119.html

The issue is that if we get an SMTPResponseException with a permanent
error code, we don't want to add all the recipients to the failures
dict, otherwise they'll all score a bounce.  That's not appropriate
because the message is being rejected by the local smtpd for reasons
having to do with the content of the message, not because there's some
delivery problem to the recipient.

Log all problems now to logs/smtp-failure.